Add eel-wasm
This is a pet project I've been working on for a while. It still has a few bugs that I'm working out, but it's functional.

I'm not sure if amateur compilers like this are appropriate for this list or not. If not, feel free not to merge this pull request. 

Thanks for the great list!